requirements.txt
streamlit==0.71.0
Pillow==8.0.1
